Improve ctags scripts so they will inform you if any package is missing

+++ bin/ctags-for-php-project.zsh
@@ -1,7 +1,21 @@
 #!/bin/zsh
 
+# Check for `ctags` package
+if ! $(command -v ctags >/dev/null 2>&1)
+then
+    echo "Install ctags first"
+    echo
+    echo -e "\tsudo apt-get install exuberant-ctags"
+    echo
+    echo "Once installed, run this script again."
+    exit 1
+fi
+
+# Remove already generated tags file (if exists)
 rm -f $(git rev-parse --show-toplevel)/.git/tags
-ctags-exuberant -f $(git rev-parse --show-toplevel)/.git/tags \
+
+# Generate tags file and store it in .git/tags
+ctags -f $(git rev-parse --show-toplevel)/.git/tags \
     -h ".inc" \
     -R \
     --exclude="*\.js" \
